Refactor RemoteBlockInputStream.
